Convert square icons to block buttons
